+++ cmd/packageloader2/loader/variables.go
@@ -1,0 +1,5 @@
+package loader
+
+var SKIP_PACKAGES = []string{"al", "gl", "glfw", "time", "os", "gltext", "cx", "json", "cipher", "tcp"}
+var FileHashMap = make(map[string]string)
+var PackageHashMap = make(map[string]string)
